Struct libatasmart::Disk

pub struct Disk {
    pub disk: PathBuf,
    // some fields omitted
}

Our ata smart disk

Fields
disk: PathBuf

The path in the filesystem to the hard drive

Methods

impl Disk

pub fn new(disk_path: &Path) -> Result<Disk, String>

This will initialize a new Disk by asking libatasmart to open it. Note that this requires root permissions usually to succeed.

pub fn get_disk_size(&mut self) -> Result<u64, String>

Returns a u64 representing the size of the disk in bytes.

pub fn check_sleep_mode(&mut self) -> Result<bool, String>

Returns a bool of true if sleep mode is supported, false otherwise.

pub fn get_power_on(&mut self) -> Result<u64, String>

Returns a u64 representing the power on time in milliseconds

pub fn get_power_cycle_count(&mut self) -> Result<u64, String>

Returns a u64 representing the number of power on cycles

pub fn get_bad_sectors(&mut self) -> Result<u64, String>

Returns a u64 representing the number of bad sections on the disk

pub fn get_temperature(&mut self) -> Result<u64, String>

Returns a u64 representing the mkelvin of the disk

pub fn get_smart_status(&mut self) -> Result<bool, String>

Returns true if the disk passed smart, false otherwise.

pub fn dump(&mut self) -> Result<(), String>

This will dump all available information to stdout about the drive

pub fn identify_is_available(&mut self) -> Result<bool, String>

pub fn smart_is_available(&mut self) -> Result<bool, String>

Query the device and return whether or not smart is supported on it

pub fn execute_smart_self_test(
&mut self,
test_type: SkSmartSelfTest
) -> Result<(), String>

pub fn smart_get_overall(&mut self) -> Result<SkSmartOverall, String>